+++ score/Calculate.java
@@ -100,7 +100,7 @@ public class Calculate {
 				recall=Calculate.confusion_matrix[i][i]/fact;
 			if(precision!=0 && recall!=0)
 				f1=2*precision*recall/(precision+recall);
-			Main.mp.appendMessage(""+(i+Cross_validation.start_num)+"\t\t"+df.format(precision)+
+			Main.mp.appendMessage(""+(int)(i+Cross_validation.start_num)+"\t\t"+df.format(precision)+
 					"\t\t"+df.format(recall)+"\t\t"+df.format(f1)+"\n");
 		}
 	}
@@ -114,13 +114,13 @@ public class Calculate {
 		Main.mp.appendMessage("\n=== Confusion Matrix ===\n\n");
 		Main.mp.appendMessage("Prediction type\n");
 		for(int i=0;i<Calculate.type;i++)
-			Main.mp.appendMessage(""+(i+Cross_validation.start_num)+"\t");
+			Main.mp.appendMessage(""+(int)(i+Cross_validation.start_num)+"\t");
 		Main.mp.appendMessage("Real type\n\n");
 		for(int i=0;i<Calculate.type;i++){
 			for(int j=0;j<Calculate.type;j++){
 				Main.mp.appendMessage(""+(int)Calculate.confusion_matrix[i][j]+"\t");
 			}
-			Main.mp.appendMessage(""+(i+Cross_validation.start_num)+"\n");
+			Main.mp.appendMessage(""+(int)(i+Cross_validation.start_num)+"\n");
 		}
 	}
